Aula2

Description

Este módulo contém definições Haskell para as tarefas proposta na Ficha 2 de LI1.

Documentation

addToList :: [Int] -> Int -> [Int]

A função addToList, recebendo uma lista de inteiros, adiciona um inteiro a cada elemento da lista.

Exemplos de Utilização

>>> addToList [1,2,3] 1
[2,3,4]

allBut :: [String] -> Char -> [String]

A função allBut, recebendo uma lista de Strings, retorna todos os elementos que não comecem por um caractere.

Exemplos de Utilização

>>> allBut ["Foo", "Bar", "Baz"] 'B'
["Foo"]

>>> allBut ["Foo", "Bar", "Baz"] 'F'
["Bar", "Baz"]

addToEveryFirst :: [(Int, Int)] -> Int -> [(Int, Int)]

A função addToEveryFirst, recebendo uma lista contendo tuplos de dois inteiros, adiciona um outro inteiro a todos os primeiros elementos de cada tuplo contido na lista.

Exemplos de Utilização

>>> addToEveryFirst [(1,2),(3,4),(5,6)] 1
[(2,2),(4,4),(6,6)]

maxSecond :: [(Int, Int)] -> Int

A função maxSecond, recebendo uma lista contendo tuplos de dois inteiros, retorna o maior valor entre todos os segundos elementos de cada tuplo.

Exemplos de Utilização

>>> maxSecond [(1,2),(3,4)]
4

getNext :: Int -> Int

A função getNext, recebendo um inteiro contido entre 0 e 9, retorna o número precedente desse inteiro. Considera-se o zero (0) como o precedente do nove (9).

Exemplos de Utilização

>>> getNext 1
2

advanceAllDigits :: [Int] -> [Int]

A função advanceAllDigits, recebendo uma lista de inteiros contidos entre 0 e 9, retorna o número precedente de cada elemento.

Considera-se o zero (0) como o precedente do nove (9).

Exemplos de Utilização

>>> advanceAllDigits [1,2,3]
[2,3,4]

advanceAllVowels :: [Char] -> [Char]

A função advanceAllVowels, recebendo uma lista de vogais, retorna a vogal precedente de cada elemento.

Considera-se o "a" como o precedente do "u".

Exemplos de Utilização

>>> advanceAllVowels ['a','e','i','o','u']
['e','i','o','u','a']

type Nome = String

type Coordenada = (Int, Int)

data Movimento

Constructors

N

S

E

W

type Movimentos = [Movimento]

data PosicaoPessoa

Constructors

Pos Nome Coordenada

posicao :: PosicaoPessoa -> Movimentos -> PosicaoPessoa

A função posicao, recebendo uma PosicaoPessoa e uma lista de Movimentos, retorna a nova posição após executada a lista de movimentos.

Exemplos de Utilização

>>> posicao (Pos "Foo" (0,0)) [N,N]
Pos "Foo" (0,2)

>>> posicao (Pos "Foo" (0,-2)) [S,S]
Pos "Foo" (0,2)

>>> posicao (Pos "Foo" (-2,0)) [W,W]
Pos "Foo" (0,2)

>>> posicao (Pos "Foo" (2,0)) [E,E]
Pos "Foo" (0,2)

posicoesM :: [PosicaoPessoa] -> Movimento -> [PosicaoPessoa]

A função posicoesM, recebendo uma lista de PosicaoPessoas e um Movimento, retorna uma lista contendo as novas posições após executado o movimento em cada uma delas.

Exemplos de Utilização

>>> posicoesM [(Pos "Foo" (0,0)),(Pos "Bar" (1,1))] N
[Pos "Foo" (0,1),Pos "Bar" (1,2)]

posicoesMs :: [PosicaoPessoa] -> Movimentos -> [PosicaoPessoa]

A função posicoesMs, recebendo uma lista de PosicaoPessoas e uma lista de Movimentos, retorna uma lista contendo as novas posições após executada a lista de movimentos em cada uma delas.

Exemplos de Utilização

>>> posicoesMs [(Pos "Foo" (0,0)),(Pos "Bar" (1,1))] [N,W]
[Pos "Foo" (-1,1),Pos "Bar" (0,2)]

pessoasNorte :: [PosicaoPessoa] -> [Nome]

A função pessoasNorte, recebendo uma lista de PosicaoPessoas, retorna uma lista contendo os nomes das posições contendo a maior coordenada y.

Exemplos de Utilização

>>> pessoasNorte [(Pos "Foo" (0,0)),(Pos "Bar" (1,1))]
["Bar"]

>>> pessoasNorte [(Pos "Foo" (100,1)),(Pos "Bar" (1,1))]
["Foo","Bar"]