Aula3

Description

Este módulo contém definições Haskell para as tarefas proposta na Ficha 4 de LI1.

Documentation

shiftRightN :: [Int] -> Int -> [Int]

A função shiftRightN desloca todos os elementos da lista n posições para a direita. Os elementos deslocados para uma posição superior ao tamanho da lista serão colocados no inicio da lista.

Exemplos de Utilização

>>> shiftRightN [1,2,3,4,5] 2
[4,5,1,2,3]

shiftLeftN :: (Eq t, Num t) => [a] -> t -> [a]

A função shiftLeftN desloca todos os elementos da lista n posições para a esquerda. Os elementos deslocados para uma posição inferior ao tamanho da lista serão colocados no fim da lista.

Exemplos de Utilização

>>> shiftLeftN [1,2,3,4,5] 2
[3,4,5,1,2]

type Matriz = [[Int]]

Representa uma Matriz matemática

Observações

Assume-se que todas as linhas têm o mesmo número de elementos

swapLines :: Matriz -> Matriz

A função swapLines troca as posições da primeira e última linha de uma matriz.

Exemplos de Utilização

>>> swapLines [[1,2],[3,4],[5,6]]
[[5,6],[3,4],[1,2]]

swapColumns :: Matriz -> Matriz

A função swapColumns troca as posições da primeira e última coluna de uma matriz.

Exemplos de Utilização

>>> swapColumns [[1,2],[3,4],[5,6]]
[[2,1],[4,3],[6,5]]

elemIndex :: Eq t => [t] -> t -> Int

A função elemIndex retorna a posição da primeira ocorrência de um elemento, caso exista. Caso contrário retorna -1.

Exemplos de Utilização

>>> elemIndex [1,2,3] 3
2

>>> elemIndex [1,2,3] 4
-1

replaceAt :: [a] -> Int -> a -> [a]

A função replaceAt substitui o elemento localizado na posição n de uma lista por um elemento especificado.

Exemplos de Utilização

>>> replaceAt "abcdefg" 3 'X'
"abcXefg"

searchMatrix :: Matriz -> Int -> (Int, Int)

A função searchMatrix retorna a posição da primeira ocorrência de um elemento especificado numa matriz, caso exista. Caso contrário retorna -1.

Exemplos de Utilização

>>> searchMatrix [[1,0],[2,3]] 3
(1,1)

>>> searchMatrix [[1,0],[2,3]] 4
(-1,-1)

replaceMatrixAt :: Matriz -> (Int, Int) -> Int -> Matriz

A função replaceMatrixAt substitui o elemento localizado na posição (l,c) de uma matriz por um elemento especificado.

Exemplos de Utilização

>>> replaceMatrixAt [[1,0],[3,4]] (0,1) 2
[[1,2],[3,4]]