Aula5

Description

Este módulo contém definições Haskell para as tarefas proposta na Ficha 5 de LI1.

Documentation

areaQuad :: Float -> Float

A função areaQuad calcula a área de um quadrado dada a medida do seu lado.

Exemplos de Utilização

>>> areaQuad 5
25

allBut :: [String] -> Char -> [String]

A função allBut, recebendo uma lista de Strings, retorna todos os elementos que não comecem por um caractere.

Exemplos de Utilização

>>> allBut ["Foo", "Bar", "Baz"] 'B'
["Foo"]

>>> allBut ["Foo", "Bar", "Baz"] 'F'
["Bar", "Baz"]

type Matriz = [[Int]]

Representa uma Matriz matemática

Observações

Assume-se que todas as linhas têm o mesmo número de elementos

swapLines :: Matriz -> Matriz

A função swapLines troca as posições da primeira e última linha de uma matriz.

Exemplos de Utilização

>>> swapLines [[1,2],[3,4],[5,6]]
[[5,6],[3,4],[1,2]]